Fix Carbon 3 rawAddUnit type

+++ config/honeypot.php
@@ -39,7 +39,7 @@ return [
      * If the form is submitted faster than this amount of seconds
      * the form submission will be considered invalid.
      */
-    'amount_of_seconds' => env('HONEYPOT_SECONDS', 3),
+    'amount_of_seconds' => (int) env('HONEYPOT_SECONDS', 3),
 
     /*
      * This class is responsible for sending a response to requests that
